Preter un livre

+++ src/main/java/repository/AdherentRepository.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 
 public interface AdherentRepository {
     List<Adherent> findAll();
-    Optional<Adherent> findById(Long id);
+    Optional<Adherent> findById(int id);
     Adherent save(Adherent adherent);
     void deleteById(Long id);
     Adherent findByEmailAndMotDePasse(String email, String motDePasse);
