+++ src/main/java/repository/BibliothecaireRepository.java
@@ -8,7 +8,7 @@ import model.Bibliothecaire;
 public interface BibliothecaireRepository {
 
     List<Bibliothecaire> findAll();
-    Optional<Bibliothecaire> findById(Long id);
+    Optional<Bibliothecaire> findById(int id);
     Bibliothecaire save(Bibliothecaire categorie);
     void deleteById(Long id);
     Bibliothecaire findByNomAndMotDePasse(String nom, String motDePasse);
